src/test/java/ru/netology/StatsServiceTest.java
package ru.netology;
import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;

class StatsServiceTest {

    @Test
    void shouldArraySum() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 180;

        long actual = service.arraySum(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldAverageSum() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 15;

        long actual = service.averageSum(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldMaxSumMonth() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 8;

        long actual = service.maxSumMonth(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldMinSumMonth() {
        StatsService service = new StatsService();
        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 9;

        long actual = service.minSumMonth(sales);
        assertEquals(expected, actual);
    }

    @Test
    void shouldMonthMaxAverage() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 5;

        long actual = service.monthMaxAverage(sales);
        assertEquals(expected, actual);
    }


    @Test
    void shouldMonthMinAverage() {
        StatsService service = new StatsService();

        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        long expected = 5;

        long actual = service.monthMinAverage(sales);
        assertEquals(expected, actual);
    }

}
